renamed pakage name for release
