+++ ANDROID-MAC-HACK-KI/app/build.gradle
@@ -8,7 +8,7 @@ android {
     compileSdkVersion 26
     buildToolsVersion "26.0.2"
     defaultConfig {
-        applicationId 'com.amigotrip.amigo'
+        applicationId 'com.amigotrip.android'
         minSdkVersion 19
         targetSdkVersion 26
         versionCode 2
