+++ ANDROID-MAC-HACK-KI/app/src/androidTest/java/com/amigotrip/android/ExampleInstrumentedTest.kt
@@ -1,12 +1,10 @@
-package com.amigotrip.amigo
+package com.amigotrip.android
 
 import android.support.test.InstrumentationRegistry
 import android.support.test.runner.AndroidJUnit4
-
+import org.junit.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
-
-import org.junit.Assert.*
 
 /**
  * Instrumented test, which will execute on an Android device.
