+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/android/NumberPickerDialog.kt
@@ -1,4 +1,4 @@
-package com.amigotrip.amigo
+package com.amigotrip.android
 
 import android.app.AlertDialog
 import android.app.Dialog
