+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/android/activities/ChooseLangActivity.kt
@@ -1,11 +1,11 @@
-package com.amigotrip.amigo.activities
+package com.amigotrip.android.activities
 
 import android.app.Activity
 import android.content.Intent
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
 import android.widget.ArrayAdapter
-import com.amigotrip.amigo.R
+import com.amigotrip.anroid.R
 import kotlinx.android.synthetic.main.activity_choose_lang.*
 import java.io.BufferedReader
 import java.io.InputStreamReader
