+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/android/activities/MessagesActivity.kt
@@ -1,13 +1,13 @@
-package com.amigotrip.amigo.activities
+package com.amigotrip.android.activities
 
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
-import com.amigotrip.amigo.R
+import com.amigotrip.anroid.R
 
-class ProfileActivity : AppCompatActivity() {
+class MessagesActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_profile)
+        setContentView(R.layout.activity_messages)
     }
 }
