+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/android/activities/SearchActivity.kt
@@ -1,13 +1,13 @@
-package com.amigotrip.amigo.activities
+package com.amigotrip.android.activities
 
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
-import com.amigotrip.amigo.R
+import com.amigotrip.anroid.R
 
-class IntroActivity : AppCompatActivity() {
+class SearchActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_intro)
+        setContentView(R.layout.activity_search)
     }
 }
