+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/android/activities/WelcomeActivity.kt
@@ -1,8 +1,8 @@
-package com.amigotrip.amigo.activities
+package com.amigotrip.android.activities
 
 import android.os.Bundle
 import android.support.v7.app.AppCompatActivity
-import com.amigotrip.amigo.R
+import com.amigotrip.anroid.R
 import kotlinx.android.synthetic.main.activity_done.*
 
 class WelcomeActivity : AppCompatActivity() {
