+++ ANDROID-MAC-HACK-KI/app/src/main/java/com/amigotrip/android/datas/Party.kt
@@ -1,4 +1,4 @@
-package com.amigotrip.amigo.datas
+package com.amigotrip.android.datas
 
 /**
  * Created by Zimincom on 2017. 10. 19..
