+++ ANDROID-MAC-HACK-KI/app/src/main/res/layout/activity_choose_lang.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.amigotrip.amigo.activities.ChooseLangActivity">
+    tools:context="com.amigotrip.android.activities.ChooseLangActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
