+++ ANDROID-MAC-HACK-KI/app/src/main/res/layout/activity_messages.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.amigotrip.amigo.activities.MessagesActivity">
+    tools:context="com.amigotrip.android.activities.MessagesActivity">
 
 </android.support.constraint.ConstraintLayout>
