+++ ANDROID-MAC-HACK-KI/app/src/main/res/layout/activity_new_party.xml
@@ -14,7 +14,7 @@
         <android.support.constraint.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context="com.amigotrip.amigo.activities.NewPartyActivity"
+            tools:context="com.amigotrip.amigo.com.amigo.android.activities.NewPartyActivity"
             tools:layout_editor_absoluteX="0dp"
             tools:layout_editor_absoluteY="81dp">
