+++ ANDROID-MAC-HACK-KI/app/src/main/res/layout/activity_search.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.amigotrip.amigo.activities.SearchActivity">
+    tools:context="com.amigotrip.android.activities.SearchActivity">
 
 </android.support.constraint.ConstraintLayout>
