+++ ANDROID-MAC-HACK-KI/app/src/main/res/layout/activity_sign_in.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.amigotrip.amigo.activities.SignInActivity"
+    tools:context="com.amigotrip.android.activities.SignInActivity"
     tools:layout_editor_absoluteX="0dp"
     tools:layout_editor_absoluteY="81dp">
 
@@ -26,7 +26,7 @@
         app:layout_constraintHorizontal_bias="0.501"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/ic_logo_green" />
+        android:src="@drawable/ic_logo_green" />
 
     <Button
         android:id="@+id/btn_view"
