+++ ANDROID-MAC-HACK-KI/app/src/test/java/com/amigotrip/android/ExampleUnitTest.kt
@@ -1,8 +1,7 @@
-package com.amigotrip.amigo
+package com.amigotrip.android
 
+import org.junit.Assert.assertEquals
 import org.junit.Test
-
-import org.junit.Assert.*
 
 /**
  * Example local unit test, which will execute on the development machine (host).
